+++ R/XcmsExperiment-functions.R
@@ -71,7 +71,7 @@
 #' @noRd
 .subset_xcms_experiment <- function(x, i = integer(),
                                     keepChromPeaks = TRUE,
-                                    keepAdjustedRtime = FALSE,
+                                    keepAdjustedRtime = TRUE,
                                     keepFeatures = FALSE,
                                     ignoreHistory = FALSE,
                                     keepSampleIndex = FALSE, ...) {